Se agregó validación de eliminar vacante mediante policy

+++ app/Livewire/MostrarVacantes.php
@@ -5,13 +5,25 @@ namespace App\Livewire;
 use App\Models\Vacante;
 use Livewire\Component;
 use Livewire\Attributes\On;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 
 class MostrarVacantes extends Component
 {
     #[On('eliminarVacante')]
 
+    
+
     public function eliminarVacante(Vacante $vacante){
-        $vacante->delete();
+        if(Gate::allows('delete', $vacante)){
+            // Elimino Imagen
+            $result = Storage::delete('public/vacantes/'.$vacante->imagen);
+             // Elimino Vacante
+            $vacante->delete();
+        }else{
+            return redirect()->route('vacantes.index');
+        }
+        
     }
 
     public function render()
